feat: isolated editor side effects
