+++ app/components/write/editor/default-editor-content.tsx
