+++ app/hooks/use-keyboard-save.ts
@@ -1,0 +1,16 @@
+import { useEffect } from 'react';
+
+export function useKeyboardSave(callback: () => void) {
+  useEffect(() => {
+    const onKeyDown = (e: KeyboardEvent) => {
+      if (e.metaKey && e.key === 's') {
+        e.preventDefault();
+        callback();
+      }
+    };
+    document.addEventListener('keydown', onKeyDown);
+    return () => {
+      document.removeEventListener('keydown', onKeyDown);
+    };
+  }, [callback]);
+}
